Reorganize LionWebRepoClientBulkApiFunctionalTest
