+++ repo-client-testing/README.md
@@ -1,0 +1,7 @@
+# Repo-client-testing
+
+Building functional tests relying on the LionWeb Repo Client is difficult as 
+it requires running a Postgres DB and the LW Repository. To make it easier
+we created this module, which is also used when testing the `repo-client` module
+in this same project. We expect it will be useful to others building applications
+based on the `repo-client` module.
